% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{add}
\alias{add}
\title{Add a package}
\usage{
add(packages, remotes = c())
}
\arguments{
\item{packages}{Packages to add}

\item{remotes}{Remotes to add}
}
\description{
Add a package
}
\examples{
\dontrun{

jetpack::add("randomForest")

jetpack::add(c("randomForest", "DBI"))

jetpack::add("DBI@1.0.0")

jetpack::add("plyr", remote="hadley/plyr")

jetpack::add("plyr", remote="local::/path/to/plyr")
}
}
