% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check that all dependencies are installed}
\usage{
check()
}
\value{
\code{TRUE} if all dependencies are installed, \code{FALSE} otherwise, invisibly
}
\description{
Check that all dependencies are installed
}
\examples{
\dontrun{

jetpack::check()
}
}
