% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli}
\alias{cli}
\title{Install the command line interface}
\usage{
cli(file = NULL)
}
\arguments{
\item{file}{The file to create}
}
\description{
Install the command line interface
}
\examples{
\dontrun{

jetpack::cli()
}
}
