% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jetpack.R
\name{info}
\alias{info}
\title{Get info for a package}
\usage{
info(package)
}
\arguments{
\item{package}{Package to get info for}
}
\description{
Get info for a package
}
\examples{
\dontrun{

jetpack::info("stringr")

jetpack::info("stringr@1.0.0")
}
}
