% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jetpack.R
\name{jetpack.add}
\alias{jetpack.add}
\title{Add a package}
\usage{
jetpack.add(packages, remotes = c())
}
\arguments{
\item{packages}{Packages to add}

\item{remotes}{Remotes to add}
}
\description{
Add a package
}
