% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jetpack.R
\name{jetpack.remove}
\alias{jetpack.remove}
\title{Remove a package}
\usage{
jetpack.remove(packages, remotes)
}
\arguments{
\item{packages}{Packages to remove}

\item{remotes}{Remotes to remove}
}
\description{
Remove a package
}
