% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jetpack.R
\name{jetpack.update}
\alias{jetpack.update}
\title{Update a package}
\usage{
jetpack.update(packages, remotes)
}
\arguments{
\item{packages}{Packages to update}

\item{remotes}{Remotes to update}
}
\description{
Update a package
}
