% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install packages for a project}
\usage{
install(deployment = FALSE)
}
\arguments{
\item{deployment}{Use deployment mode}
}
\description{
Install packages for a project
}
\examples{
\dontrun{

jetpack::install()
}
}
