% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove}
\alias{remove}
\title{Remove a package}
\usage{
remove(packages, remotes = c())
}
\arguments{
\item{packages}{Packages to remove}

\item{remotes}{Remotes to remove}
}
\value{
No return value
}
\description{
Remove a package
}
\examples{
\dontrun{

jetpack::remove("randomForest")

jetpack::remove(c("randomForest", "DBI"))

jetpack::remove("plyr", remote="hadley/plyr")
}
}
