% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{update}
\alias{update}
\title{Update a package}
\usage{
update(packages, remotes = c())
}
\arguments{
\item{packages}{Packages to update}

\item{remotes}{Remotes to update}
}
\description{
Update a package
}
\examples{
\dontrun{

jetpack::update("randomForest")

jetpack::update(c("randomForest", "DBI"))
}
}
